Ex16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
 
 int main(void)
 {
-	unsigned long Fibo1 = 1;
-	unsigned long n = 0;
-	unsigned long Fibo2 = 1;
+	float Fibo1 = 1;
+	float n = 0;
+	float Fibo2 = 1;
 	int m = 0;
 
 	while (m < 98)
@@ -18,9 +18,9 @@ int main(void)
 		Fibo2 = Fibo2 + Fibo1;
 		Fibo1 = n;
 		if (m < 97)
-			printf("%ld, ", Fibo1);
+			printf("%f, ", Fibo1);
 		if (m == 97)
-			printf("%ld", Fibo1);
+			printf("%f", Fibo1);
 		m++;
 	}
 	printf("\n");
